Create WPS_forum_dump
With the aim of describing the WPS db scheme re: WPS forums, this is a dump of a succession of queries, followed with var_dump( $wpdb->last_query, $result );

By default WPS has 3 forum categories, and one topic with two replies.  I've created a new instance of WPS, added two topics, to populate all categories, and made one category child of another.

I've also created a group.  In WPS, groups can have a forum, with no category hierarchy, so they are not listed in the table symposium_cats.  This group is private, its forum contains one topic with a reply.